[InventorySystem] Refactored bad namespace naming...
