+++ src/Rhisis.World/Game/Chat/CreateItemCommand.cs
@@ -3,7 +3,7 @@ using Rhisis.Core.IO;
 using Rhisis.Core.Structures.Game;
 using Rhisis.World.Game.Entities;
 using Rhisis.World.Systems;
-using Rhisis.World.Systems.Events.Statistics;
+using Rhisis.World.Systems.Events.Inventory;
 
 namespace Rhisis.World.Game.Chat
 {
