+++ src/Rhisis.World/Handlers/InventoryHandler.cs
@@ -3,7 +3,7 @@ using Rhisis.Core.Network;
 using Rhisis.Core.Network.Packets;
 using Rhisis.Core.Network.Packets.World;
 using Rhisis.World.Systems;
-using Rhisis.World.Systems.Events.Statistics;
+using Rhisis.World.Systems.Events.Inventory;
 
 namespace Rhisis.World.Handlers
 {
