+++ src/Rhisis.World/Systems/Events/Inventory/InventoryCreateItemEventArgs.cs
@@ -1,4 +1,4 @@
-﻿namespace Rhisis.World.Systems.Events.Statistics
+﻿namespace Rhisis.World.Systems.Events.Inventory
 {
     public class InventoryCreateItemEventArgs : InventoryEventArgs
     {
