+++ src/Rhisis.World/Systems/InventorySystem.cs
@@ -7,14 +7,13 @@ using Rhisis.World.Game.Core;
 using Rhisis.World.Game.Core.Interfaces;
 using Rhisis.World.Game.Entities;
 using Rhisis.World.Game.Structures;
-using Rhisis.World.Handlers;
-using Rhisis.World.Systems.Events.Statistics;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using Rhisis.Core.Structures.Game;
 using Rhisis.World.Packets;
+using Rhisis.World.Systems.Events.Inventory;
 
 namespace Rhisis.World.Systems
 {
