+++ src/Rhisis.World/Systems/VisibilitySystem.cs
@@ -2,10 +2,9 @@
 using Rhisis.World.Game.Core;
 using Rhisis.World.Game.Core.Interfaces;
 using Rhisis.World.Game.Entities;
-using Rhisis.World.Handlers;
 using System;
 using System.Linq.Expressions;
-using Rhisis.Core.IO;
+using Rhisis.World.Packets;
 
 namespace Rhisis.World.Systems
 {
